security(luks): passphrase via stdin, piano crypted mai su disco + rename config→pathDefaults
La passphrase LUKS non tocca più il filesystem: viene passata via stdin
a cryptsetup (--key-file -) e il piano JSON in modalità --crypted viene
inviato a oa via pipe invece che scritto su file.
Rinominato pkg/config in pkg/pathDefaults per liberare il nome per una
futura TUI di configurazione.
