+++ coa/pkg/planner/create-exclude-list.go
@@ -1,7 +1,7 @@
 package planner
 
 import (
-	"coa/pkg/config"
+	"coa/pkg/pathDefaults"
 	"os"
 	"strings"
 )
@@ -9,7 +9,7 @@ import (
 // GenerateExcludeList crea il file .list dinamico per mksquashfs.
 // La 'G' maiuscola permette a remaster.go di chiamarla liberamente.
 func GenerateExcludeList(mode string, isGitHubAction bool) string {
-	outPath := config.ExcludeList
+	outPath := pathDefaults.ExcludeList
 	var excludes []string
 
 	// ==========================================================
@@ -115,7 +115,7 @@ func GenerateExcludeList(mode string, isGitHubAction bool) string {
 		}
 	}
 
-	os.MkdirAll(config.StagingDir, 0755)
+	os.MkdirAll(pathDefaults.StagingDir, 0755)
 
 	// Uniamo tutto con a capo
 	fileContent := strings.Join(excludes, "\n") + "\n"
